ajustes

+++ src/main/java/br/com/escola/biblioteca/service/EditoraService.java
@@ -5,6 +5,7 @@ import br.com.escola.biblioteca.dto.EditoraResponseDTO;
 import br.com.escola.biblioteca.entity.Editora;
 import br.com.escola.biblioteca.exception.NotFoundException;
 import br.com.escola.biblioteca.repository.EditoraRepository;
+import br.com.escola.biblioteca.repository.LivroRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +17,9 @@ public class EditoraService {
 
     @Autowired
     private EditoraRepository editoraRepository;
+
+    @Autowired
+    private LivroRepository livroRepository;
 
     public EditoraResponseDTO criar(EditoraRequestDTO dto) {
         validarCnpj(dto.cnpj());
@@ -73,6 +77,11 @@ public class EditoraService {
         if (!editoraRepository.existsById(id)) {
             throw new NotFoundException("Editora não encontrada com id: " + id);
         }
+
+        if (!livroRepository.findByEditoraId(id).isEmpty()) {
+            throw new RuntimeException(
+                    "Não é possível excluir editora com livros cadastrados. Delete os livros primeiro.");
+        }
         editoraRepository.deleteById(id);
     }
 
@@ -87,7 +96,7 @@ public class EditoraService {
         }
         String cnpjNumeros = cnpj.replaceAll("[.\\-/]", "");
         if (cnpjNumeros.length() != 14) {
-            throw new RuntimeException("CNPJ inválido. Deve conter 14 dígitos. Exemplo: 12.345.678/0001-90.");
+            throw new RuntimeException("CNPJ inválido. Deve conter 14 dígitos. Exemplo: 12345678000190.");
         }
         if (!cnpjNumeros.matches("\\d+")) {
             throw new RuntimeException("CNPJ deve conter apenas números.");
